+++ app/models/basket_item.rb
@@ -4,6 +4,6 @@ class BasketItem < ApplicationRecord
   belongs_to :item
   belongs_to :basket
 
-  attribute :price, :integer, default: 0
-  attribute :final_price, :integer, default: 0
+  attribute :price, :float, default: 0.0
+  attribute :final_price, :float, default: 0.0
 end
